ShardMetricAggregator moved to new file
