+++ core/src/main/resources/filodb-defaults.conf
@@ -160,7 +160,7 @@ filodb {
   # spread-assignment = [ { _ws_ = demo, _ns_ = App-Test001, _spread_ = 5 } ]
   spread-assignment = []
 
-  shard-key-level-ingestion-metrics-enabled = true
+  shard-key-level-ingestion-metrics-enabled = true  // TODO(a_theimer): reference this
   # info config used for metric breakdown
   cluster-type = "raw" // possible values: downsample, raw, recRule, aggregates etc.
   # Name of the deployment partition that this FiloDB cluster belongs to
